backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
groq==0.4.0
yt-dlp==2024.1.7
beautifulsoup4==4.12.2
requests==2.31.0
googlesearch-python==1.2.3
python-dotenv==1.0.0
lxml==4.9.3